Factor dash logic out of pattern_bucket_features
