+++ src/data/bucket/pattern_bucket_features.ts
@@ -27,13 +27,6 @@ export function hasPattern(type: string, layers: PatternStyleLayers, options: Po
             patterns[constantPattern.from] =  true;
         }
 
-        // Also check for data-driven dasharray
-        if (type === 'line') {
-            const dasharrayProperty = (layer.paint as PossiblyEvaluated<any, any>).get('line-dasharray');
-            if (dasharrayProperty && !dasharrayProperty.isConstant()) {
-                hasPattern = true;
-            }
-        }
     }
 
     return hasPattern;
@@ -42,7 +35,6 @@ export function hasPattern(type: string, layers: PatternStyleLayers, options: Po
 export function addPatternDependencies(type: string, layers: PatternStyleLayers, patternFeature: BucketFeature, parameters: { zoom: number; globalState: Record<string, any> }, options: PopulateParameters) {
     const {zoom, globalState} = parameters;
     const patterns = options.patternDependencies;
-    const dashes = options.dashDependencies;
 
     for (const layer of layers) {
         const patternProperty = (layer.paint  as PossiblyEvaluated<any, any>).get(`${type}-pattern`);
@@ -63,29 +55,6 @@ export function addPatternDependencies(type: string, layers: PatternStyleLayers,
             // save for layout
             patternFeature.patterns[layer.id] = {min, mid, max};
         }
-
-        // Handle data-driven dasharray
-        if (type === 'line') {
-            const dasharrayProperty = (layer.paint as PossiblyEvaluated<any, any>).get('line-dasharray');
-            if (dasharrayProperty && dasharrayProperty.value.kind !== 'constant') {
-                const round = false;
-
-                const min = {dasharray: dasharrayProperty.value.evaluate({zoom: zoom - 1}, patternFeature, {}), round};
-                const mid = {dasharray: dasharrayProperty.value.evaluate({zoom}, patternFeature, {}), round};
-                const max = {dasharray: dasharrayProperty.value.evaluate({zoom: zoom + 1}, patternFeature, {}), round};
-
-                const minKey = JSON.stringify(min);
-                const midKey = JSON.stringify(mid);
-                const maxKey = JSON.stringify(max);
-
-                dashes[minKey] = min;
-                dashes[midKey] = mid;
-                dashes[maxKey] = max;
-
-                patternFeature.dashes[layer.id] = {min: minKey, mid: midKey, max: maxKey};
-            }
-        }
     }
     return patternFeature;
 }
-
